I think it's ready (enough.)
The functionality is there, even if there are things I'd still like to
implement you can use it and Joy code that only uses words that exist
runs on it (see gcd.)

I still have to write docs, but it feels right now like I'll be writing
about this for months at least, if not years, so I'm letting the code out
there early, before the documentation is written.

One of the things that I like about this is that, if you admit that the
GUI Text window is a form of output, this script is a quine.