Trace via coroutine-ish.
Instead of appending the stack and expression each time through we yield
them and pass them back in.  This allows for "injecting" different values
for the stack and/or expression during a run, because we have, in addition
to the current stack and expression about to be executed, a continuation
of the run with the whole (Python) call stack including any currently
running combinators higher up from the current execution "frame".

We should be able to do wacky things.  For example, we can change the
implementation of a function halfway through a run (by assigning a
different callable to the FunctionWrapper's f attribute) and see what
effect that has on the execution. We could substitute a Python function
for a definition and see how that affects runtime.