The combinators work with the new hottness now.
The ones that still recurse into joy() throw away changes to the
dictionary made "inside" those calls so we get a kind of lexical scoping
for free.  The combinators that work by expression rewriting don't
actually do any processing so there is no "inner" scope for a mutant
dictionary to exist within, and so any changes made to the D by some
as-yet-hypothetical "inscribe" command would be, in effect, global.

The joy() ones could be written to return the dictionary from the call
and in fact there are a few possibilities to explore there.